# Disable firebase performance plugin to speed up build time