--- a/gnd/build.gradle
+++ b/gnd/build.gradle
@@ -68,6 +68,9 @@ android {
         }
         debug {
             resValue("string", "GOOGLE_MAPS_API_KEY", googleMapsApiKey)
+            FirebasePerformance {
+                instrumentationEnabled false
+            }
         }
     }