Fixes #460 

https://firebase.google.com/docs/perf-mon/disable-sdk?platform=android#disable-gradle-plugin

Disables the plugin by setting `instrumentationEnabled` to `false` only for debug build type.
This prevents `gnd:transformClassesWithFirebasePerformancePluginForDebug` task from getting executed.

To verify the changes, I ran 3 tests and performed the following steps for each.
 - Disabled gradle cache from `gradle.properties` so that it does not affect test results
 - Clean project
 - Build project

Tests summary:
1. No changes(master branch): 57s
2. Commented out `apply plugin: 'com.google.firebase.firebase-perf'`: 23s
3. Set `instrumentationEnabled` as false: 22s

(with caching enabled, it only takes 5-7s with the new changes)

Full output:
----------------
## Test 1
1. Clean project
2. Build project

```
Executing tasks: [:gnd:assembleDebug] in project /Users/ashobhit/Code/ground-android

Configuration on demand is an incubating feature.

> Configure project :gnd
WARNING: The following project options are deprecated and have been removed: 
android.enableUnitTestBinaryResources
The raw resource for unit test functionality is removed.


WARNING: The option setting 'android.enableSeparateAnnotationProcessing=true' is experimental and unsupported.
The current default is 'false'.


> Configure project :android-gmaps-addons
WARNING: The following project options are deprecated and have been removed: 
android.enableUnitTestBinaryResources
The raw resource for unit test functionality is removed.



> Task :android-gmaps-addons:preBuild UP-TO-DATE
> Task :android-gmaps-addons:preDebugBuild UP-TO-DATE
> Task :android-gmaps-addons:generateDebugBuildConfig
> Task :android-gmaps-addons:generateDebugResValues
> Task :gnd:preBuild UP-TO-DATE
> Task :gnd:preDebugBuild UP-TO-DATE
> Task :android-gmaps-addons:compileDebugAidl NO-SOURCE
> Task :android-gmaps-addons:compileDebugRenderscript NO-SOURCE
> Task :android-gmaps-addons:generateDebugResources
> Task :gnd:mergeDebugShaders
> Task :gnd:compileDebugAidl NO-SOURCE
> Task :android-gmaps-addons:packageDebugResources
> Task :gnd:compileDebugShaders
> Task :gnd:generateDebugAssets

> Task :gnd:processDebugGoogleServices
Parsing json file: /Users/ashobhit/Code/ground-android/gnd/src/debug/google-services.json

> Task :gnd:mainApkListPersistenceDebug
> Task :gnd:createDebugCompatibleScreenManifests
> Task :gnd:extractDeepLinksDebug
> Task :android-gmaps-addons:packageDebugRenderscript NO-SOURCE
> Task :gnd:generateDebugBuildConfig
> Task :android-gmaps-addons:mergeDebugShaders
> Task :gnd:compileDebugRenderscript NO-SOURCE
> Task :android-gmaps-addons:compileDebugShaders
> Task :android-gmaps-addons:generateDebugAssets
> Task :gnd:writeDebugApplicationId
> Task :android-gmaps-addons:packageDebugAssets
> Task :android-gmaps-addons:extractDeepLinksDebug
> Task :android-gmaps-addons:javaPreCompileDebug
> Task :android-gmaps-addons:parseDebugLocalResources
> Task :android-gmaps-addons:processDebugManifest
> Task :gnd:generateSafeArgsDebug
> Task :gnd:mergeDebugAssets
> Task :android-gmaps-addons:generateDebugRFile
> Task :android-gmaps-addons:processDebugAnnotationsWithJavac SKIPPED
> Task :android-gmaps-addons:compileDebugJavaWithJavac
> Task :android-gmaps-addons:bundleLibCompileDebug
> Task :android-gmaps-addons:compileDebugLibraryResources
> Task :android-gmaps-addons:processDebugJavaRes NO-SOURCE
> Task :android-gmaps-addons:bundleLibResDebug

> Task :gnd:processDebugManifest
/Users/ashobhit/Code/ground-android/gnd/src/main/AndroidManifest.xml:24:9-31:50 Warning:
	activity#com.google.firebase.auth.internal.FederatedSignInActivity@android:launchMode was tagged at AndroidManifest.xml:24 to replace other declarations but no other declaration present

> Task :gnd:fabricGenerateResourcesDebug
> Task :gnd:prepareLintJar UP-TO-DATE
> Task :gnd:prepareLintJarForPublish UP-TO-DATE
> Task :gnd:generateDebugSources
> Task :android-gmaps-addons:bundleLibRuntimeDebug
> Task :gnd:dataBindingExportBuildInfoDebug
> Task :android-gmaps-addons:mergeDebugJniLibFolders
> Task :gnd:generateDebugResValues
> Task :gnd:generateDebugResources
> Task :android-gmaps-addons:mergeDebugNativeLibs
> Task :android-gmaps-addons:stripDebugDebugSymbols
> Task :android-gmaps-addons:copyDebugJniLibsProjectOnly
> Task :gnd:mergeDebugResources
> Task :gnd:processDebugResources
> Task :gnd:processDebugJavaRes NO-SOURCE
> Task :gnd:checkDebugDuplicateClasses
> Task :gnd:mergeDebugJniLibFolders
> Task :gnd:dataBindingMergeGenClassesDebug
> Task :gnd:validateSigningDebug
> Task :gnd:dataBindingExportFeaturePackageIdsDebug
> Task :gnd:javaPreCompileDebug
> Task :gnd:dataBindingMergeDependencyArtifactsDebug
> Task :gnd:dataBindingGenBaseClassesDebug

> Task :gnd:processDebugAnnotationsWithJavac

> Task :gnd:compileDebugJavaWithJavac
ANTLR Tool version 4.5.3 used for code generation does not match the current runtime version 4.7.1ANTLR Runtime version 4.5.3 used for parser compilation does not match the current runtime version 4.7.1ANTLR Tool version 4.5.3 used for code generation does not match the current runtime version 4.7.1ANTLR Runtime version 4.5.3 used for parser compilation does not match the current runtime version 4.7.1Note: Some input files use or override a deprecated API.
Note: Recompile with -Xlint:deprecation for details.
Note: Some input files use unchecked or unsafe operations.
Note: Recompile with -Xlint:unchecked for details.

> Task :gnd:compileDebugSources

> Task :gnd:transformClassesWithFirebasePerformancePluginForDebug
java.lang.ClassNotFoundException: java.util.concurrent.Flow$Publisher
java.lang.ClassNotFoundException: java.util.concurrent.Flow$Publisher
java.lang.ClassNotFoundException: java.util.concurrent.Flow$Processor
java.lang.ClassNotFoundException: java.util.concurrent.Flow$Processor
java.lang.ClassNotFoundException: java.util.concurrent.Flow$Subscriber
java.lang.ClassNotFoundException: java.util.concurrent.Flow$Subscriber

> Task :gnd:mergeDebugNativeLibs
> Task :gnd:mergeDebugJavaResource
> Task :gnd:dexBuilderDebug
> Task :gnd:stripDebugDebugSymbols
> Task :gnd:multiDexListDebug
> Task :gnd:mergeDexDebug
> Task :gnd:packageDebug
> Task :gnd:assembleDebug

BUILD SUCCESSFUL in 57s
56 actionable tasks: 54 executed, 2 up-to-date
```


## Test 2
1. Comment out `apply plugin: 'com.google.firebase.firebase-perf'`
2. Clean project
3. Build project

```
Executing tasks: [:gnd:assembleDebug] in project /Users/ashobhit/Code/ground-android

Configuration on demand is an incubating feature.

> Configure project :gnd
WARNING: The following project options are deprecated and have been removed: 
android.enableUnitTestBinaryResources
The raw resource for unit test functionality is removed.


WARNING: The option setting 'android.enableSeparateAnnotationProcessing=true' is experimental and unsupported.
The current default is 'false'.


> Configure project :android-gmaps-addons
WARNING: The following project options are deprecated and have been removed: 
android.enableUnitTestBinaryResources
The raw resource for unit test functionality is removed.



> Task :android-gmaps-addons:preBuild UP-TO-DATE
> Task :android-gmaps-addons:preDebugBuild UP-TO-DATE
> Task :android-gmaps-addons:generateDebugBuildConfig
> Task :android-gmaps-addons:generateDebugResValues
> Task :gnd:preBuild UP-TO-DATE
> Task :gnd:preDebugBuild UP-TO-DATE
> Task :android-gmaps-addons:compileDebugAidl NO-SOURCE
> Task :android-gmaps-addons:compileDebugRenderscript NO-SOURCE
> Task :android-gmaps-addons:generateDebugResources
> Task :android-gmaps-addons:packageDebugResources
> Task :android-gmaps-addons:processDebugManifest
> Task :gnd:mergeDebugShaders
> Task :android-gmaps-addons:packageDebugRenderscript NO-SOURCE
> Task :gnd:compileDebugShaders
> Task :gnd:generateDebugAssets
> Task :android-gmaps-addons:mergeDebugShaders

> Task :gnd:processDebugGoogleServices
Parsing json file: /Users/ashobhit/Code/ground-android/gnd/src/debug/google-services.json

> Task :android-gmaps-addons:compileDebugShaders
> Task :android-gmaps-addons:javaPreCompileDebug
> Task :gnd:compileDebugAidl NO-SOURCE
> Task :android-gmaps-addons:generateDebugAssets
> Task :gnd:compileDebugRenderscript NO-SOURCE
> Task :android-gmaps-addons:packageDebugAssets
> Task :android-gmaps-addons:parseDebugLocalResources
> Task :gnd:mergeDebugAssets
> Task :android-gmaps-addons:extractDeepLinksDebug
> Task :gnd:mainApkListPersistenceDebug
> Task :gnd:createDebugCompatibleScreenManifests
> Task :gnd:extractDeepLinksDebug
> Task :android-gmaps-addons:generateDebugRFile
> Task :android-gmaps-addons:processDebugAnnotationsWithJavac SKIPPED
> Task :android-gmaps-addons:compileDebugJavaWithJavac
> Task :android-gmaps-addons:bundleLibCompileDebug
> Task :android-gmaps-addons:compileDebugLibraryResources
> Task :android-gmaps-addons:processDebugJavaRes NO-SOURCE
> Task :android-gmaps-addons:bundleLibResDebug
> Task :android-gmaps-addons:bundleLibRuntimeDebug
> Task :android-gmaps-addons:mergeDebugJniLibFolders

> Task :gnd:processDebugManifest
/Users/ashobhit/Code/ground-android/gnd/src/main/AndroidManifest.xml:24:9-31:50 Warning:
	activity#com.google.firebase.auth.internal.FederatedSignInActivity@android:launchMode was tagged at AndroidManifest.xml:24 to replace other declarations but no other declaration present

> Task :android-gmaps-addons:mergeDebugNativeLibs
> Task :android-gmaps-addons:stripDebugDebugSymbols
> Task :android-gmaps-addons:copyDebugJniLibsProjectOnly
> Task :gnd:fabricGenerateResourcesDebug
> Task :gnd:generateDebugBuildConfig
> Task :gnd:writeDebugApplicationId
> Task :gnd:generateSafeArgsDebug
> Task :gnd:prepareLintJar UP-TO-DATE
> Task :gnd:prepareLintJarForPublish UP-TO-DATE
> Task :gnd:generateDebugSources
> Task :gnd:dataBindingExportBuildInfoDebug
> Task :gnd:generateDebugResValues
> Task :gnd:generateDebugResources
> Task :gnd:mergeDebugResources
> Task :gnd:processDebugResources
> Task :gnd:processDebugJavaRes NO-SOURCE
> Task :gnd:dataBindingExportFeaturePackageIdsDebug
> Task :gnd:dataBindingMergeGenClassesDebug
> Task :gnd:javaPreCompileDebug
> Task :gnd:dataBindingMergeDependencyArtifactsDebug
> Task :gnd:dataBindingGenBaseClassesDebug

> Task :gnd:processDebugAnnotationsWithJavac

> Task :gnd:compileDebugJavaWithJavac
ANTLR Tool version 4.5.3 used for code generation does not match the current runtime version 4.7.1ANTLR Runtime version 4.5.3 used for parser compilation does not match the current runtime version 4.7.1ANTLR Tool version 4.5.3 used for code generation does not match the current runtime version 4.7.1ANTLR Runtime version 4.5.3 used for parser compilation does not match the current runtime version 4.7.1Note: Some input files use or override a deprecated API.
Note: Recompile with -Xlint:deprecation for details.
Note: Some input files use unchecked or unsafe operations.
Note: Recompile with -Xlint:unchecked for details.

> Task :gnd:compileDebugSources
> Task :gnd:checkDebugDuplicateClasses
> Task :gnd:desugarDebugFileDependencies
> Task :gnd:dexBuilderDebug
> Task :gnd:mergeDebugJniLibFolders
> Task :gnd:validateSigningDebug
> Task :gnd:mergeDebugJavaResource
> Task :gnd:mergeDebugNativeLibs
> Task :gnd:stripDebugDebugSymbols
> Task :gnd:multiDexListDebug
> Task :gnd:mergeDexDebug
> Task :gnd:packageDebug
> Task :gnd:assembleDebug

BUILD SUCCESSFUL in 23s
56 actionable tasks: 54 executed, 2 up-to-date
```

## Test 3
1. Set `instrumentationEnabled` as false in `build.gradle`
2. Clean project 
3. Build project

```
Executing tasks: [:gnd:assembleDebug] in project /Users/ashobhit/Code/ground-android

Configuration on demand is an incubating feature.

> Configure project :gnd
WARNING: The following project options are deprecated and have been removed: 
android.enableUnitTestBinaryResources
The raw resource for unit test functionality is removed.


WARNING: The option setting 'android.enableSeparateAnnotationProcessing=true' is experimental and unsupported.
The current default is 'false'.


> Configure project :android-gmaps-addons
WARNING: The following project options are deprecated and have been removed: 
android.enableUnitTestBinaryResources
The raw resource for unit test functionality is removed.



> Task :android-gmaps-addons:preBuild UP-TO-DATE
> Task :android-gmaps-addons:preDebugBuild UP-TO-DATE
> Task :android-gmaps-addons:generateDebugBuildConfig
> Task :android-gmaps-addons:generateDebugResValues
> Task :android-gmaps-addons:processDebugManifest
> Task :android-gmaps-addons:compileDebugAidl NO-SOURCE
> Task :android-gmaps-addons:compileDebugRenderscript NO-SOURCE
> Task :android-gmaps-addons:generateDebugResources
> Task :android-gmaps-addons:packageDebugResources
> Task :gnd:preBuild UP-TO-DATE
> Task :gnd:preDebugBuild UP-TO-DATE
> Task :android-gmaps-addons:javaPreCompileDebug
> Task :android-gmaps-addons:packageDebugRenderscript NO-SOURCE
> Task :android-gmaps-addons:parseDebugLocalResources
> Task :gnd:mergeDebugShaders
> Task :gnd:compileDebugShaders
> Task :gnd:generateDebugAssets

> Task :gnd:processDebugGoogleServices
Parsing json file: /Users/ashobhit/Code/ground-android/gnd/src/debug/google-services.json

> Task :gnd:compileDebugAidl NO-SOURCE
> Task :gnd:compileDebugRenderscript NO-SOURCE
> Task :android-gmaps-addons:mergeDebugShaders
> Task :android-gmaps-addons:compileDebugShaders
> Task :android-gmaps-addons:generateDebugAssets
> Task :gnd:mainApkListPersistenceDebug
> Task :android-gmaps-addons:packageDebugAssets
> Task :android-gmaps-addons:extractDeepLinksDebug
> Task :gnd:createDebugCompatibleScreenManifests
> Task :gnd:mergeDebugAssets
> Task :gnd:extractDeepLinksDebug
> Task :android-gmaps-addons:generateDebugRFile
> Task :android-gmaps-addons:processDebugAnnotationsWithJavac SKIPPED
> Task :android-gmaps-addons:compileDebugJavaWithJavac
> Task :android-gmaps-addons:bundleLibCompileDebug
> Task :android-gmaps-addons:compileDebugLibraryResources
> Task :android-gmaps-addons:processDebugJavaRes NO-SOURCE
> Task :android-gmaps-addons:bundleLibResDebug
> Task :android-gmaps-addons:bundleLibRuntimeDebug
> Task :android-gmaps-addons:mergeDebugJniLibFolders
> Task :android-gmaps-addons:mergeDebugNativeLibs
> Task :android-gmaps-addons:stripDebugDebugSymbols
> Task :android-gmaps-addons:copyDebugJniLibsProjectOnly

> Task :gnd:processDebugManifest
/Users/ashobhit/Code/ground-android/gnd/src/main/AndroidManifest.xml:24:9-31:50 Warning:
	activity#com.google.firebase.auth.internal.FederatedSignInActivity@android:launchMode was tagged at AndroidManifest.xml:24 to replace other declarations but no other declaration present

> Task :gnd:fabricGenerateResourcesDebug
> Task :gnd:generateDebugBuildConfig
> Task :gnd:writeDebugApplicationId
> Task :gnd:generateSafeArgsDebug
> Task :gnd:prepareLintJar UP-TO-DATE
> Task :gnd:prepareLintJarForPublish UP-TO-DATE
> Task :gnd:generateDebugSources
> Task :gnd:dataBindingExportBuildInfoDebug
> Task :gnd:generateDebugResValues
> Task :gnd:generateDebugResources
> Task :gnd:dataBindingMergeDependencyArtifactsDebug
> Task :gnd:mergeDebugResources
> Task :gnd:processDebugResources
> Task :gnd:processDebugJavaRes NO-SOURCE
> Task :gnd:checkDebugDuplicateClasses
> Task :gnd:mergeDebugJniLibFolders
> Task :gnd:validateSigningDebug
> Task :gnd:javaPreCompileDebug
> Task :gnd:dataBindingMergeGenClassesDebug
> Task :gnd:dataBindingGenBaseClassesDebug
> Task :gnd:dataBindingExportFeaturePackageIdsDebug

> Task :gnd:processDebugAnnotationsWithJavac

> Task :gnd:compileDebugJavaWithJavac
ANTLR Tool version 4.5.3 used for code generation does not match the current runtime version 4.7.1ANTLR Runtime version 4.5.3 used for parser compilation does not match the current runtime version 4.7.1ANTLR Tool version 4.5.3 used for code generation does not match the current runtime version 4.7.1ANTLR Runtime version 4.5.3 used for parser compilation does not match the current runtime version 4.7.1Note: Some input files use or override a deprecated API.
Note: Recompile with -Xlint:deprecation for details.
Note: Some input files use unchecked or unsafe operations.
Note: Recompile with -Xlint:unchecked for details.

> Task :gnd:compileDebugSources
> Task :gnd:dexBuilderDebug
> Task :gnd:mergeDebugNativeLibs
> Task :gnd:stripDebugDebugSymbols
> Task :gnd:mergeDebugJavaResource
> Task :gnd:multiDexListDebug
> Task :gnd:mergeDexDebug
> Task :gnd:packageDebug
> Task :gnd:assembleDebug

BUILD SUCCESSFUL in 22s
55 actionable tasks: 53 executed, 2 up-to-date
```